работа на стадии завершения
