+++ app.py
@@ -1,7 +1,9 @@
 import logging
 
 from flask import Flask
-from main.main_views import main_blueprint, tag_blueprint, search_blueprint, bookmarks_blueprint
+from main.main_views import main_blueprint, search_blueprint, tag_blueprint
+from bookmark.bokkmark_views import bookmarks_blueprint
+from api.api_views import api_blueprint
 
 app = Flask(__name__)
 
@@ -9,8 +11,10 @@ app.register_blueprint(main_blueprint)
 app.register_blueprint(tag_blueprint)
 app.register_blueprint(search_blueprint)
 app.register_blueprint(bookmarks_blueprint)
+app.register_blueprint(api_blueprint)
 
-logging.basicConfig(filename="logs.log", encoding="utf-8", level=logging.INFO)
+logging.basicConfig(filename="logs.log", encoding="utf-8", level=logging.INFO,
+                    format='%(asctime)s [%(levelname)s] %(message)s')
 
 if __name__ == "__main__":
     app.run()
