+++ bookmark/templates/all_bookmarks.html
@@ -14,7 +14,7 @@
 
 <header class="container header">
     <div class="header__wrapper">
-        <a href="/" class="header__btn-back">
+        <a href="/static" class="header__btn-back">
             <svg class="header__btn-icon" width="19" height="16">
                 <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#back"></use>
             </svg>
