+++ main/templates/all_posts.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>SKYPROGRAM</title>
-    <link href="../static/css/styles.min.css" rel="stylesheet">
+    <link href="../../static/css/styles.min.css" rel="stylesheet">
 
 </head>
 
@@ -15,7 +15,7 @@
 <header class="container header">
     <div class="header__wrapper">
         <a href="/search" class="header__btn-back">
-            <img src="../static/img/glass.png" alt="" class="header__loupe-icon" width="28">
+            <img src="../../static/img/glass.png" alt="" class="header__loupe-icon" width="28">
             <span class="header__loupe-counter"></span>
         </a>
         <a class="header__logo-link">
@@ -24,7 +24,7 @@
             </svg>
         </a>
         <a href="/all_bookmarks" class="header__bookmark-btn">
-            <img src="../static/img/bookmark.png" alt="" class="header__bookmark-icon" width="28">
+            <img src="../../static/img/bookmark.png" alt="" class="header__bookmark-icon" width="28">
             <span class="header__bookmark-counter">{{ len_bookmark }}</span>
         </a>
     </div>
@@ -56,7 +56,7 @@
                     {{ post.content[:50] }}...</p>
                 <div class="item__post-info">
                     <div class="item__counter">
-                        <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+                        <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
                         <span class="item__counter-text">{{ post.views_count }}</span>
                     </div>
                 </div>
@@ -77,7 +77,7 @@
                                 <input type="submit" value="Добавить в закладки" class="button">
                             </div>
                             <a href="/add_bookmarks" class="item__btn-more">
-                                <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+                                <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
                             </a>
                         </form>
                     </button>
