+++ main/templates/search.html
@@ -11,7 +11,7 @@
 <body>
 <header class="container header">
     <div class="header__wrapper">
-        <a href="/" class="header__btn-back">
+        <a href="/static" class="header__btn-back">
             <svg class="header__btn-icon" width="19" height="16">
                 <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#back"></use>
             </svg>
@@ -58,7 +58,7 @@
                 </div>
             </div>
             <div class="item__post-bottom">
-                <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+                <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
                 <button type="button" class="item__bookmark-btn">
                     <form action="/bookmarks/add_bookmarks" method="POST" enctype="multipart/form-data">
                             <div>
@@ -74,7 +74,7 @@
                                 <input type="submit" value="Добавить в закладки" class="button">
                             </div>
                             <a href="/add_bookmarks" class="item__btn-more">
-                                <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+                                <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
                             </a>
                         </form>
                 </button>
