+++ old/bookmarks.html
@@ -5,13 +5,13 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>SKYPROGRAM bookmarks</title>
-  <link href="../../static/css/styles.min.css" rel="stylesheet">
+  <link href="../static/css/styles.min.css" rel="stylesheet">
 </head>
 
 <body>
   <header class="container header">
     <div class="header__wrapper">
-      <a href="/" class="header__btn-back">
+      <a href="/static" class="header__btn-back">
         <svg class="header__btn-icon" width="19" height="16">
           <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#back"></use>
         </svg>
@@ -29,26 +29,26 @@
         <div class="item__top">
           <a href="/templates/user-feed.html-feed.html" class="item__head">
             <div class="item__ava-wrap">
-              <img src="../../static/img/ava1.png" alt="" class="item__ava" width="47">
+              <img src="../static/img/ava1.png" alt="" class="item__ava" width="47">
             </div>
             <span class="item__username">skypro.08</span>
           </a>
           <a href="#" class="item__delete">Удалить</a>
         </div>
-        <a href="/main/templates/post.html/post.html" class="item__post-link">
-          <img src="../../static/img/post5.jpg" alt="" width="348" class="item__post-img">
+        <a href="/old/post.html/post.html" class="item__post-link">
+          <img src="../static/img/post5.jpg" alt="" width="348" class="item__post-img">
         </a>
-        <p class="item__descr">Текст описания длинный... <a href="/main/templates/post.html/post.html" class="item__btn-more">Еще</a></p>
+        <p class="item__descr">Текст описания длинный... <a href="/old/post.html/post.html" class="item__btn-more">Еще</a></p>
         <div class="item__post-info">
           <div class="item__counter">
-            <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+            <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
             <span class="item__counter-text">212</span>
           </div>
         </div>
         <div class="item__post-bottom">
-          <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+          <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
           <button type="button" class="item__bookmark-btn">
-            <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+            <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
           </button>
         </div>
       </div>
@@ -56,26 +56,26 @@
         <div class="item__top">
           <a href="/templates/user-feed.html-feed.html" class="item__head">
             <div class="item__ava-wrap">
-              <img src="../../static/img/ava2.png" alt="" class="item__ava" width="47">
+              <img src="../static/img/ava2.png" alt="" class="item__ava" width="47">
             </div>
             <span class="item__username">Username</span>
           </a>
           <a href="#" class="item__delete">Удалить</a>
         </div>
-        <a href="/main/templates/post.html/post.html" class="item__post-link">
-          <img src="../../static/img/post6.jpg" alt="" width="348" class="item__post-img">
+        <a href="/old/post.html/post.html" class="item__post-link">
+          <img src="../static/img/post6.jpg" alt="" width="348" class="item__post-img">
         </a>
-        <p class="item__descr">Текст описания длинный... <a href="/main/templates/post.html/post.html" class="item__btn-more">Еще</a></p>
+        <p class="item__descr">Текст описания длинный... <a href="/old/post.html/post.html" class="item__btn-more">Еще</a></p>
         <div class="item__post-info">
           <div class="item__counter">
-            <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+            <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
             <span class="item__counter-text">212</span>
           </div>
         </div>
         <div class="item__post-bottom">
-          <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+          <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
           <button type="button" class="item__bookmark-btn">
-            <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+            <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
           </button>
         </div>
       </div>
@@ -83,26 +83,26 @@
         <div class="item__top">
           <a href="/templates/user-feed.html-feed.html" class="item__head">
             <div class="item__ava-wrap">
-              <img src="../../static/img/ava2.png" alt="" class="item__ava" width="47">
+              <img src="../static/img/ava2.png" alt="" class="item__ava" width="47">
             </div>
             <span class="item__username">Username</span>
           </a>
           <a href="#" class="item__delete">Удалить</a>
         </div>
-        <a href="/main/templates/post.html/post.html" class="item__post-link">
-          <img src="../../static/img/post7.jpg" alt="" width="348" class="item__post-img">
+        <a href="/old/post.html/post.html" class="item__post-link">
+          <img src="../static/img/post7.jpg" alt="" width="348" class="item__post-img">
         </a>
-        <p class="item__descr">Текст описания длинный... <a href="/main/templates/post.html/post.html" class="item__btn-more">Еще</a></p>
+        <p class="item__descr">Текст описания длинный... <a href="/old/post.html/post.html" class="item__btn-more">Еще</a></p>
         <div class="item__post-info">
           <div class="item__counter">
-            <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+            <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
             <span class="item__counter-text">1023</span>
           </div>
         </div>
         <div class="item__post-bottom">
-          <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+          <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
           <button type="button" class="item__bookmark-btn active">
-            <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+            <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
           </button>
         </div>
       </div>
