+++ old/index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>SKYPROGRAM</title>
-  <link href="../../static/css/styles.min.css" rel="stylesheet">
+  <link href="../static/css/styles.min.css" rel="stylesheet">
 </head>
 
 <body>
@@ -32,25 +32,25 @@
         <div class="item__top">
           <a href="/templates/user-feed.html-feed.html" class="item__head">
             <div class="item__ava-wrap">
-              <img src="../../static/img/ava1.png" alt="" class="item__ava" width="47">
+              <img src="../static/img/ava1.png" alt="" class="item__ava" width="47">
             </div>
             <span class="item__username">Johnny</span>
           </a>
         </div>
-        <a href="/main/templates/post.html/post.html" class="item__post-link">
-          <img src="../../static/img/post1.jpg" alt="" width="348" class="item__post-img">
+        <a href="/old/post.html/post.html" class="item__post-link">
+          <img src="../static/img/post1.jpg" alt="" width="348" class="item__post-img">
         </a>
         <p class="item__descr">Текст описания коротки...</p>
         <div class="item__post-info">
           <div class="item__counter">
-            <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+            <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
             <span class="item__counter-text">212</span>
           </div>
         </div>
         <div class="item__post-bottom">
-          <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+          <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
           <button type="button" class="item__bookmark-btn">
-            <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+            <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
           </button>
         </div>
       </div>
@@ -66,25 +66,25 @@
         <div class="item__top">
           <a href="/templates/user-feed.html-feed.html" class="item__head">
             <div class="item__ava-wrap">
-              <img src="../../static/img/ava2.png" alt="" class="item__ava" width="47">
+              <img src="../static/img/ava2.png" alt="" class="item__ava" width="47">
             </div>
             <span class="item__username">Username</span>
           </a>
         </div>
-        <a href="/main/templates/post.html/post.html" class="item__post-link">
-          <img src="../../static/img/post2.jpg" alt="" width="348" class="item__post-img">
+        <a href="/old/post.html/post.html" class="item__post-link">
+          <img src="../static/img/post2.jpg" alt="" width="348" class="item__post-img">
         </a>
         <p class="item__descr">Текст описания коротки...</p>
         <div class="item__post-info">
           <div class="item__counter">
-            <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+            <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
             <span class="item__counter-text">1023</span>
           </div>
         </div>
         <div class="item__post-bottom">
-          <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+          <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
           <button type="button" class="item__bookmark-btn active">
-            <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+            <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
           </button>
         </div>
       </div>
@@ -92,25 +92,25 @@
         <div class="item__top">
           <a href="/templates/user-feed.html-feed.html" class="item__head">
             <div class="item__ava-wrap">
-              <img src="../../static/img/ava2.png" alt="" class="item__ava" width="47">
+              <img src="../static/img/ava2.png" alt="" class="item__ava" width="47">
             </div>
             <span class="item__username">Username</span>
           </a>
         </div>
-        <a href="/main/templates/post.html/post.html" class="item__post-link">
-          <img src="../../static/img/post3.jpg" alt="" width="348" class="item__post-img">
+        <a href="/old/post.html/post.html" class="item__post-link">
+          <img src="../static/img/post3.jpg" alt="" width="348" class="item__post-img">
         </a>
         <p class="item__descr">Текст описания коротки...</p>
         <div class="item__post-info">
           <div class="item__counter">
-            <img src="../../static/img/eye.png" alt="" class="item__counter-icon" width="26">
+            <img src="../static/img/eye.png" alt="" class="item__counter-icon" width="26">
             <span class="item__counter-text">1023</span>
           </div>
         </div>
         <div class="item__post-bottom">
-          <a href="/main/templates/post.html/post.html" class="item__btn-more">Подробнее</a>
+          <a href="/old/post.html/post.html" class="item__btn-more">Подробнее</a>
           <button type="button" class="item__bookmark-btn">
-            <img src="../../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
+            <img src="../static/img/bookmark.png" alt="" class="item__bookmark-icon" width="24">
           </button>
         </div>
       </div>
